libtinyaes: Enable and fix style
JIRA: RTOS-783
